Add new example.vcf.gz file for vignette
Add a new example.vcf.gz file for use in the vignette, different from
the unit testing files. This file is a merged VCF file from three
different sources: from the cell lines HCT116, HKE3 and RKO, from the
data in Fasterius et al. (2017). It is simplified in that it does not
contain variant calling-specific command history, mergin history, etc,
and only contains a subset of variants from chromosome 12, in order to
keep the file size down.