Merge compare_overlaps with overlap_profiles
Merge the compare_overlaps function with the overlap_profiles function
and rename the new function to "compare_profiles", since there is no
reason to perform an overlap of two SNV profiles without also comparing
their genotypes. In other words, it was redundant to have these two
functions separated, and it saves time and thought for the user to
simply perform both at the same time. Change documentation, function
names, examples, NAMESPACE, manuals, and so on to reflect this change.